# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Movyra API (Product Contract)
  version: 0.2.0
  description: >
    Frontend-first contract aligned to DB schema (V1 + V3 + V4).
    Multi-tenant via X-Tenant header for tenant-scoped routes.
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/tenants:
    get:
      tags: [Tenants]
      summary: List tenants (master)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Sort"
        - name: q
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Paged list of tenants
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PagedTenant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Tenants]
      summary: Create tenant (master)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/tenants/{tenantId}:
    get:
      tags: [Tenants]
      summary: Get tenant by id (master)
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      tags: [Tenants]
      summary: Update tenant (master)
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/tenants/{tenantId}/deactivate:
    post:
      tags: [Tenants]
      summary: Deactivate tenant (soft delete)
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
      responses:
        "204":
          description: Deactivated
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # =========================
  # TENANT-SCOPED (requires X-Tenant)
  # =========================
...
  /api/appointments:
    get:
      tags: [Appointments]
      summary: List appointments (tenant)
      parameters:
        - $ref: "#/components/parameters/XTenant"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Sort"
        - name: from
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/AppointmentStatus" }
        - name: professionalId
          in: query
          required: false
          schema: { $ref: "#/components/schemas/Uuid" }
      responses:
        "200":
          description: Paged list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PagedAppointment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Appointments]
      summary: Create appointment (tenant)
      parameters:
        - $ref: "#/components/parameters/XTenant"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppointmentCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "409":
          description: Conflict (e.g. overlap, invalid state transition)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /api/portfolio-items:
    get:
      tags: [Portfolio]
      summary: List portfolio items (tenant)
      parameters:
        - $ref: "#/components/parameters/XTenant"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: Paged list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PagedPortfolioItem" }
    post:
      tags: [Portfolio]
      summary: Create portfolio item (tenant)
      parameters:
        - $ref: "#/components/parameters/XTenant"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PortfolioItemCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PortfolioItem" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XTenant:
      name: X-Tenant
      in: header
      required: true
      schema: { $ref: "#/components/schemas/Uuid" }
      description: Tenant id (UUID) for tenant-scoped routes.
    TenantIdPath:
      name: tenantId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Uuid" }
...
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
    Sort:
      name: sort
      in: query
      required: false
      schema: { type: string, default: "createdAt,desc" }
      description: Sort string like "createdAt,desc"
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
  schemas:
...
    Uuid: { type: string, format: uuid }
    IsoDateTime: { type: string, format: date-time }
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        traceId: { type: string, nullable: true }
    ValidationError:
      type: object
      required: [code, message, details]
      properties:
        code: { type: string, example: "VALIDATION_ERROR" }
        message: { type: string, example: "Invalid request" }
        details:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    PageMeta:
      type: object
      required: [page, size, totalItems, totalPages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        totalItems: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }

    # ---- tenants (V1)
    BillingStatus:
      type: string
      enum: [trial, active, past_due, cancelled]
    TenantCreate:
      type: object
      required: [slug, name, timezone]
      properties:
        slug:
          type: string
          minLength: 3
          maxLength: 80
          description: Unique slug (DB unique)
        name: { type: string, minLength: 2, maxLength: 200 }
        phone: { type: string, nullable: true }
        addressLine1: { type: string, nullable: true }
        addressLine2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
        zip: { type: string, nullable: true }
        country:
          type: string
          minLength: 2
          maxLength: 2
          default: "US"
        timezone:
          type: string
          default: "America/New_York"
        logoUrl: { type: string, nullable: true }
        billingEmail: { type: string, format: email, nullable: true }
        billingName: { type: string, nullable: true }
    TenantUpdate:
      type: object
      properties:
        slug: { type: string, minLength: 3, maxLength: 80 }
        name: { type: string, minLength: 2, maxLength: 200 }
        phone: { type: string, nullable: true }
        addressLine1: { type: string, nullable: true }
        addressLine2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
        zip: { type: string, nullable: true }
        country: { type: string, minLength: 2, maxLength: 2 }
        timezone: { type: string }
        logoUrl: { type: string, nullable: true }
        billingEmail: { type: string, format: email, nullable: true }
        billingName: { type: string, nullable: true }
        billingStatus: { $ref: "#/components/schemas/BillingStatus" }
        isActive: { type: boolean }
    Tenant:
      type: object
      required: [id, slug, name, country, timezone, billingStatus, isActive, createdAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        slug: { type: string }
        name: { type: string }
        phone: { type: string, nullable: true }
        addressLine1: { type: string, nullable: true }
        addressLine2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
        zip: { type: string, nullable: true }
        country: { type: string }
        timezone: { type: string }
        logoUrl: { type: string, nullable: true }
        billingEmail: { type: string, nullable: true }
        billingName: { type: string, nullable: true }
        billingStatus: { $ref: "#/components/schemas/BillingStatus" }
        isActive: { type: boolean }
        createdAt: { $ref: "#/components/schemas/IsoDateTime" }
    PagedTenant:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Tenant" }
        page: { $ref: "#/components/schemas/PageMeta" }

    # ---- customers (V1 + V3 change: phone not unique)
...
    AppointmentStatus:
      type: string
      enum: [requested, confirmed, cancelled, completed, no_show]
    AppointmentCreate:
      type: object
      required: [tenantServiceId, startAt, endAt, status ]
      properties:
        customerId: { $ref: "#/components/schemas/Uuid", nullable: true }
        professionalId: { $ref: "#/components/schemas/Uuid", nullable: true }
        tenantServiceId: { $ref: "#/components/schemas/Uuid" }
        startAt: { $ref: "#/components/schemas/IsoDateTime" }
        endAt: { $ref: "#/components/schemas/IsoDateTime" }
        status: { $ref: "#/components/schemas/AppointmentStatus" }
        priceCents: { type: integer, nullable: true }
        notes: { type: string, nullable: true }
        createdByUserId: { $ref: "#/components/schemas/Uuid", nullable: true }
...
    Appointment:
      type: object
      required: [ id, tenantId, tenantServiceId, startAt, endAt, status, createdAt ]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        tenantId: { $ref: "#/components/schemas/Uuid" }
        customerId: { $ref: "#/components/schemas/Uuid", nullable: true }
        professionalId: { $ref: "#/components/schemas/Uuid", nullable: true }
        tenantServiceId: { $ref: "#/components/schemas/Uuid" }
        startAt: { $ref: "#/components/schemas/IsoDateTime" }
        endAt: { $ref: "#/components/schemas/IsoDateTime" }
        status: { $ref: "#/components/schemas/AppointmentStatus" }
        priceCents: { type: integer, nullable: true }
        notes: { type: string, nullable: true }
        createdByUserId: { $ref: "#/components/schemas/Uuid", nullable: true }
        createdAt: { $ref: "#/components/schemas/IsoDateTime" }

        # optional DTO sugar for UI (not stored in DB)
        serviceSnapshot:
          type: object
          nullable: true
          properties:
            code: { type: string }
            name: { type: string }
            durationMinutes: { type: integer }
            priceCents: { type: integer }
    PagedAppointment:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Appointment" }
        page: { $ref: "#/components/schemas/PageMeta" }

    # ---- service catalog (V1 + i18n)
...
    PortfolioItemCreate:
      type: object
      required: [ title, imageUrl ]
      properties:
        title: { type: string, minLength: 2, maxLength: 200 }
        description: { type: string, nullable: true }
        imageUrl: { type: string }
        sortOrder: { type: integer, default: 100 }
        isActive: { type: boolean, default: true }
...
    PortfolioItem:
      type: object
      required: [ id, tenantId, title, imageUrl, sortOrder, isActive, createdAt ]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        tenantId: { $ref: "#/components/schemas/Uuid" }
        title: { type: string }
        description: { type: string, nullable: true }
        imageUrl: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
        createdAt: { $ref: "#/components/schemas/IsoDateTime" }
    PagedPortfolioItem:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/PortfolioItem" }
        page: { $ref: "#/components/schemas/PageMeta" }